requirements.txt
asgiref==3.8.1
Django==5.1.4
djangorestframework==3.15.2
gunicorn==23.0.0
iniconfig==2.0.0
packaging==24.2
pluggy==1.5.0
sqlparse==0.5.3
psycopg2-binary==2.9.10